Adding _DisplayKey type
Instead of passing around untyped tuples for cache keys, lets have
a clearly typed object for this.

This makes for more readable code, and additionally corrects the
data model statement of intent that some cache keys should be displayed
as "dim", instead informing the frontend about whether the cache key
is "strict" or not, allowing the frontend to decide how to display
a strict or non-strict key.

This patch does the following:

  * types.py: Add _DisplayKey

  * element.py: Return a _DisplayKey from Element._get_display_key()

  * Other sources: Updated to use the display key object